togehter with model execution tested

+++ src/model_executor/ModelExecutor.cpp
@@ -181,11 +181,11 @@ void ModelExecutor::setModelInput(Result<torch::executor::Method>& method, std::
     
 }
 
-int ModelExecutor::getNumberOfInputValues(Result<torch::executor::Method>& method){
+unsigned int ModelExecutor::getNumberOfInputValues(Result<torch::executor::Method>& method){
     const torch::executor::EValue input_new = method->get_input(0);
     Tensor te = input_new.payload.as_tensor;
 
-    return te.numel();
+    return static_cast<unsigned int>(te.numel());
 }
 
 void ModelExecutor::printModelInput(Result<torch::executor::Method>& method)
